refactor(core): Extract ExecuteContext out of NodeExecutionFunctions (no-changelog) (#11853)
