+++ packages/core/src/node-execution-context/__tests__/execute-single-context.test.ts
@@ -12,15 +12,11 @@ import type {
 	Expression,
 	INodeType,
 	INodeTypes,
-	OnError,
-	ContextType,
-	IContextObject,
 	ICredentialDataDecryptedObject,
-	ISourceData,
-	ITaskMetadata,
 } from 'n8n-workflow';
-import { ApplicationError, NodeHelpers } from 'n8n-workflow';
+import { ApplicationError } from 'n8n-workflow';
 
+import { describeCommonTests } from './shared-tests';
 import { ExecuteSingleContext } from '../execute-single-context';
 
 describe('ExecuteSingleContext', () => {
@@ -45,6 +41,7 @@ describe('ExecuteSingleContext', () => {
 	const expression = mock<Expression>();
 	const workflow = mock<Workflow>({ expression, nodeTypes });
 	const node = mock<INode>({
+		name: 'Test Node',
 		credentials: {
 			[testCredentialType]: {
 				id: 'testCredentialId',
@@ -58,7 +55,7 @@ describe('ExecuteSingleContext', () => {
 	const additionalData = mock<IWorkflowExecuteAdditionalData>({ credentialsHelper });
 	const mode: WorkflowExecuteMode = 'manual';
 	const runExecutionData = mock<IRunExecutionData>();
-	const connectionInputData = mock<INodeExecutionData[]>();
+	const connectionInputData: INodeExecutionData[] = [];
 	const inputData: ITaskDataConnections = { main: [[{ json: { test: 'data' } }]] };
 	const executeData = mock<IExecuteData>();
 	const runIndex = 0;
@@ -84,81 +81,12 @@ describe('ExecuteSingleContext', () => {
 		expression.getParameterValue.mockImplementation((value) => value);
 	});
 
-	describe('getExecutionCancelSignal', () => {
-		it('should return the abort signal', () => {
-			expect(executeSingleContext.getExecutionCancelSignal()).toBe(abortSignal);
-		});
-	});
-
-	describe('continueOnFail', () => {
-		afterEach(() => {
-			node.onError = undefined;
-			node.continueOnFail = false;
-		});
-
-		it('should return false for nodes by default', () => {
-			expect(executeSingleContext.continueOnFail()).toEqual(false);
-		});
-
-		it('should return true if node has continueOnFail set to true', () => {
-			node.continueOnFail = true;
-			expect(executeSingleContext.continueOnFail()).toEqual(true);
-		});
-
-		test.each([
-			['continueRegularOutput', true],
-			['continueErrorOutput', true],
-			['stopWorkflow', false],
-		])('if node has onError set to %s, it should return %s', (onError, expected) => {
-			node.onError = onError as OnError;
-			expect(executeSingleContext.continueOnFail()).toEqual(expected);
-		});
-	});
-
-	describe('evaluateExpression', () => {
-		it('should evaluate the expression correctly', () => {
-			const expression = '$json.test';
-			const expectedResult = 'data';
-			const resolveSimpleParameterValueSpy = jest.spyOn(
-				workflow.expression,
-				'resolveSimpleParameterValue',
-			);
-			resolveSimpleParameterValueSpy.mockReturnValue(expectedResult);
-
-			expect(executeSingleContext.evaluateExpression(expression, itemIndex)).toEqual(
-				expectedResult,
-			);
-
-			expect(resolveSimpleParameterValueSpy).toHaveBeenCalledWith(
-				`=${expression}`,
-				{},
-				runExecutionData,
-				runIndex,
-				itemIndex,
-				node.name,
-				connectionInputData,
-				mode,
-				expect.objectContaining({}),
-				executeData,
-			);
-
-			resolveSimpleParameterValueSpy.mockRestore();
-		});
-	});
-
-	describe('getContext', () => {
-		it('should return the context object', () => {
-			const contextType: ContextType = 'node';
-			const expectedContext = mock<IContextObject>();
-			const getContextSpy = jest.spyOn(NodeHelpers, 'getContext');
-			getContextSpy.mockReturnValue(expectedContext);
-
-			expect(executeSingleContext.getContext(contextType)).toEqual(expectedContext);
-
-			expect(getContextSpy).toHaveBeenCalledWith(runExecutionData, contextType, node);
-
-			getContextSpy.mockRestore();
-		});
+	describeCommonTests(executeSingleContext, {
+		abortSignal,
+		node,
+		workflow,
+		executeData,
+		runExecutionData,
 	});
 
 	describe('getInputData', () => {
@@ -264,56 +192,6 @@ describe('ExecuteSingleContext', () => {
 				'last',
 				'params',
 			]);
-		});
-	});
-
-	describe('getInputSourceData', () => {
-		it('should return the input source data correctly', () => {
-			const inputSourceData = mock<ISourceData>();
-			executeData.source = { main: [inputSourceData] };
-
-			expect(executeSingleContext.getInputSourceData()).toEqual(inputSourceData);
-		});
-
-		it('should throw an error if the source data is missing', () => {
-			executeData.source = null;
-
-			expect(() => executeSingleContext.getInputSourceData()).toThrow(ApplicationError);
-		});
-	});
-
-	describe('logAiEvent', () => {
-		it('should log the AI event correctly', () => {
-			const eventName = 'ai-tool-called';
-			const msg = 'test message';
-
-			executeSingleContext.logAiEvent(eventName, msg);
-
-			expect(additionalData.logAiEvent).toHaveBeenCalledWith(eventName, {
-				executionId: additionalData.executionId,
-				nodeName: node.name,
-				workflowName: workflow.name,
-				nodeType: node.type,
-				workflowId: workflow.id,
-				msg,
-			});
-		});
-	});
-
-	describe('setMetadata', () => {
-		it('sets metadata on execution data', () => {
-			const metadata: ITaskMetadata = {
-				subExecution: {
-					workflowId: '123',
-					executionId: 'xyz',
-				},
-			};
-
-			expect(executeSingleContext.getExecuteData().metadata?.subExecution).toEqual(undefined);
-			executeSingleContext.setMetadata(metadata);
-			expect(executeSingleContext.getExecuteData().metadata?.subExecution).toEqual(
-				metadata.subExecution,
-			);
 		});
 	});
 });
