+++ packages/core/src/node-execution-context/index.ts
@@ -1,4 +1,5 @@
 // eslint-disable-next-line import/no-cycle
+export { ExecuteContext } from './execute-context';
 export { ExecuteSingleContext } from './execute-single-context';
 export { HookContext } from './hook-context';
 export { LoadOptionsContext } from './load-options-context';
@@ -6,3 +7,5 @@ export { PollContext } from './poll-context';
 export { SupplyDataContext } from './supply-data-context';
 export { TriggerContext } from './trigger-context';
 export { WebhookContext } from './webhook-context';
+
+export { getAdditionalKeys } from './utils';
